Verify Python version when running (must be 3.6 or above)
